pyproject.toml
[project]
name = "sure-newsletter"
version = "0.1.0"
description = "Beautiful Streamlit newsletter platform"
requires-python = ">=3.8"
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "streamlit-extras>=0.3.5",
    "streamlit-lottie>=0.0.5",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.0",
]
